Split js library tests out of test_other.py into their own file.  NFC
The `test_other.py` is over 15,000 lines long so its to to carve out
larger subsections where possible.
