+++ test/jslib/test_jslib_custom_settings.js
